clean up groq files
