+++ main-site/lib/getTags.ts
@@ -16,7 +16,7 @@ import {
 } from 'lodash-es';
 import { parseDate } from '~/lib/date';
 import { Fact, Tag } from '~/lib/groq/fact.partial.groq';
-import { OrderOfIssues } from '~/lib/groq/orderOfIssues.groq';
+import { OrderOfIssues } from '~/lib/groq/orderOfIssues.projection';
 
 export function sortTags(tags: Tag[], orderOfIssues: OrderOfIssues) {
   const noDupes = uniqBy(
