+++ main-site/lib/groq/README.md
@@ -1,0 +1,1 @@
+This folder contains Sanity GROQ projections and their type definitions
