--- main-site/lib/groq/factPageCeleb.groq.ts
@@ -1,10 +1,0 @@
-import groq from 'groq';
-import { Celeb, celebPartialGroq } from '~/lib/groq/celeb.partial.groq';
-
-export type FactPageCelebGroq = Celeb;
-
-export const factPageCelebGroq = groq`
-  *[_type == 'celeb' && slug.current == $slug][0]{
-    ${celebPartialGroq}
-  }
-`;
