+++ main-site/lib/groq/orderOfIssues.projection.ts
@@ -1,0 +1,5 @@
+import groq from 'groq';
+
+export type OrderOfIssues = string[];
+
+export const orderOfIssuesProjection = groq`topics[]->{name}.name`;
